requirements.txt
pandas
numpy
python-dotenv #read the environment variables.evn
mysql-connector-python
pymysql
scikit-learn